Added in the ability to make a basic Product Estimation Sheet
Okay, so now the script will allow for a Cost Production Estimate Excel Workbook to be passed in. If passed in, the Metrics Extractor will attempt to create a simple estimate for how much 1 job would cost.

This is a minimal proof of concept, and I will have to work with Hook Hua to improve upon my calculations, as he is the creator of the Cost Production Estimate workbooks.